app_web/requirements.txt
hydralit_components==1.0.6
matplotlib==3.6.2
numpy==1.22.4
pandas==1.4.4
streamlit==1.15.0
Pillow==9.3.0

oauth2client==4.1.3
oauthlib==3.2.2
gspread==5.9.0
streamlit-aggrid==0.3.4.post3